refactor(oauth): move the pool rotation kernel out of the Codex namespace
